+++ packages/ng-primitives/internal/src/index.ts
@@ -9,7 +9,6 @@ export {
   provideExitAnimationManager,
 } from './exit-animation/exit-animation-manager';
 export * from './interactions/index';
-export { safeTakeUntilDestroyed } from './observables/take-until-destroyed';
 export * from './signals/explicit-effect';
 export * from './signals/sync-state';
 export * from './style-injector/style-injector';
