+++ packages/ng-primitives/utils/src/index.ts
@@ -5,3 +5,4 @@ export { injectDisposables } from './helpers/disposables';
 export { uniqueId } from './helpers/unique-id';
 export { onBooleanChange, onChange } from './signals';
 export { injectDimensions } from './ui/dimensions';
+export { safeTakeUntilDestroyed } from './observables/take-until-destroyed';
